+++ docs/conf.py
@@ -187,7 +187,7 @@ sitemap_excludes = [
 ################################
 
 html_static_path = ["_static"]
-# templates_path = ["_templates"]
+templates_path = ["_templates"]
 
 #############
 # Redirects #
@@ -380,13 +380,14 @@ exclude_patterns = [
 ]
 
 # Adds custom CSS files, located remotely or in 'html_static_path'.
-# html_css_files = [
-#     "https://assets.ubuntu.com/v1/d86746ef-cookie_banner.css",
-# ]
+html_css_files = [
+    "cookie-banner.css",
+]
 
 # Adds custom JavaScript files, located remotely or in 'html_static_path'.
 html_js_files = [
     "js/overwrite_links.js",
+    "js/bundle.js",
 ]
 
 # Global substitutions generated from versions.env, so product names and